#63 Layout 컴포넌트의 세부 컴포넌트로 분리하고 React Reducer 를 별도 파일로 분리하기
