+++ src/App.js
@@ -1,15 +1,16 @@
 import React, {useEffect} from "react";
 import {HashRouter, Route, Switch} from "react-router-dom";
 import './App.less';
-import AppLayout from "./components/AppLayout";
-import Login from "./components/Login";
+import AppLayout from "./components/layout/AppLayout";
+import Login from "./components/login/Login";
 import AuthRoute from "./components/router/AuthRoute";
 import {adminConfig} from "./config/admin.config";
 import {AuthService} from "./auth/auth.service";
 import {Modal} from "antd";
 import {EventBroadcaster, SHOW_ERROR_MESSAGE_EVENT_TOPIC} from "./event/event.broadcaster";
 import {ExclamationCircleOutlined} from "@ant-design/icons";
-import GoogleOAuthRedirect from "./components/GoogleOAuthRedirect";
+import {LayoutProvider} from "./components/layout/AppLayoutContext";
+import OAuthLoginResult from "./components/login/OAuthLoginResult";
 
 const App = () => {
   useEffect(() => {
@@ -33,7 +34,7 @@ const App = () => {
         cancelButtonProps: {style: {display: 'none'}},
       });
     });
-  }, );
+  },);
 
   const existQueryStringInUrl = (url) => {
     return url.split("?")[1] ? true : false;
@@ -41,24 +42,26 @@ const App = () => {
 
   const getQueryStringInUrl = (url) => {
     if (existQueryStringInUrl(url)) {
-     return url.split("?")[1];
+      return url.split("?")[1];
     } else {
       return null;
     }
   }
 
   return (<>
-    <HashRouter>
-      <Switch>
-        <Route path={adminConfig.authentication.loginUrl} component={Login}/>
-        <Route path={adminConfig.authentication.googleOAuthRedirectLoginUrl} component={GoogleOAuthRedirect}/>
-        {adminConfig.authentication.used ? (
-          <AuthRoute path={'/'} component={AppLayout}/>
-        ) : (
-          <Route path={'/'} component={AppLayout}/>
-        )}
-      </Switch>
-    </HashRouter>
+    <LayoutProvider>
+      <HashRouter>
+        <Switch>
+          <Route path={adminConfig.authentication.loginUrl} component={Login}/>
+          <Route path={adminConfig.authentication.oauthLoginResultUrl} component={OAuthLoginResult}/>
+          {adminConfig.authentication.used ? (
+            <AuthRoute path={'/'} component={AppLayout}/>
+          ) : (
+            <Route path={'/'} component={AppLayout}/>
+          )}
+        </Switch>
+      </HashRouter>
+    </LayoutProvider>
   </>)
 };
