+++ src/components/layout/Footer.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Layout} from "antd";
+
+function Footer () {
+  return (
+    <Layout.Footer style={{textAlign: "center"}}>
+      better ADMIN ©2021 Created by bettercode
+    </Layout.Footer>
+  )
+}
+
+export default Footer;
